# perf: avoid copy