try avoid copy on matrix

before
```
[2m2025-04-11T06:44:51.537413Z[0m [32m INFO[0m [2mtest_io_middle_param::test[0m[2m:[0m Time to obfuscate: 288.818003042s
[2m2025-04-11T06:48:54.146015Z[0m [32m INFO[0m [2mtest_io_middle_param::test[0m[2m:[0m Time for evaluation: 242.611432292s
[2m2025-04-11T06:48:54.146017Z[0m [32m INFO[0m [2mtest_io_middle_param::test[0m[2m:[0m Total time: 531.429435334s

Initial physical memory: 17.78 MB
Final physical memory: 736.61 MB
Peak physical memory: 2.22 GB
Total physical memory increase: 718.83 MB
```

after
```
2025-04-11T07:54:30.220939Z  INFO test_io_middle_param::test: Time to obfuscate: 317.842486667s
2025-04-11T07:58:48.581925Z  INFO test_io_middle_param::test: Time for evaluation: 258.356424291s
2025-04-11T07:58:48.581927Z  INFO test_io_middle_param::test: Total time: 576.198910958s

===== SUMMARY =====
Initial physical memory: 17.92 MB
Final physical memory: 592.12 MB
Peak physical memory: 1.92 GB
Total physical memory increase: 574.20 MB
```